UI modifed

+++ app/controllers/EmployeesController.php
@@ -10,7 +10,7 @@ class EmployeesController extends \BaseController
         $shops = Shop::all();
 
         $item = new Item();
-        return View::make('employees.index', compact('employees', 'item', 'categories','shops'));
+        return View::make('employees.index', compact('employees', 'item', 'categories', 'shops'));
     }
 
     public function create()
@@ -35,94 +35,18 @@ class EmployeesController extends \BaseController
         }
 
         if (Input::has('updateEmployee')) {
+
+            $validator_for_update = Validator::make($data = Input::all(), Employee::$update_rules);
+            if ($validator_for_update->fails()) {
+                return Redirect::back()->withErrors($validator_for_update)->withInput();
+            }
+
             $e = Employee::where('name', Input::get('name'))->first();
             Employee::destroy($e->id);
             $data['name'] = $data['new_name'];
             Employee::create($data);
         }
 
-        /* Category */
-        if (Input::has('createCategory')) Category::create($data);
-
-        if (Input::has('deleteCategory')) {
-            $c = Category::where('Bumon', Input::get('Bumon'))->first();
-            Category::destroy($c->id);
-        }
-
-        if (Input::has('updateCategory')) {
-            $c = Category::where('Bumon', Input::get('Bumon'))->first();
-            Category::destroy($c->id);
-            $data['Bumon'] = $data['new_categoryName'];
-            Employee::create($data);
-        }
-
-         /* Shop */
-        if (Input::has('createShop')) Shop::create($data);
-
-        if (Input::has('deleteShop')) {
-            $s = Shop::where('Tenpo', Input::get('Tenpo'))->first();
-            Shop::destroy($s->id);
-        }
-
-        if (Input::has('updateShop')) {
-            $s = Shop::where('Tenpo', Input::get('Tenpo'))->first();
-            Shop::destroy($s->id);
-            $data['Tenpo'] = $data['new_shopName'];
-            Shop::create($data);
-        }
-
-
-
-        /* Item */
-        if (Input::has('createItem')) {
-
-            $id = DB::table('BTSMAS')->count();
-            $id++;
-            $file = $id . '.png';
-
-            if (move_uploaded_file($_FILES['upload']['tmp_name'], $file)) {
-                echo $file;
-            } else {
-                echo "error";
-            }
-
-            $data['contents'] = file_get_contents($file);
-            unlink($file);
-            Item::create($data);
-        }
-
-        if (Input::has('deleteItem')) {
-            $id = Input::get('idItem');
-            Item::destroy($id);
-        }
-
-        if (Input::has('updateItem')) {
-            $id = Input::get('idItem');
-            $item = Item::find($id);
-            $item->title = Input::get('title');
-            $item->price = Input::get('price');
-            $item->genka = Input::get('genka');
-            $item->Bumon = Input::get('Bumon');
-            $item->Kosu  = Input::get('Kosu');
-            if (file_exists($_FILES['upload']['tmp_name']))
-            {
-                $file = $id . '.png';
-                move_uploaded_file($_FILES['upload']['tmp_name'], $file);
-                $item->contents = file_get_contents($file);
-                unlink($file);
-            }
-            $item->save();
-        }
-
-        if (Input::get('selectedItem')) {
-            $targetID = Input::get('selectedItem');
-            $item = Item::find($targetID);
-            $employees = Employee::all();
-
-            $shops = Shop::all();
-            $categories = Category::all();
-            return View::make('employees.index', compact('employees', 'item', 'categories','shops'));
-        }
 
         return Redirect::route('employees.index');
     }
